Live node swapping, improving handling unexpected exception

+++ src/zedwallet++/ParseArguments.h
@@ -18,7 +18,7 @@ struct Config
     std::string host;
     
     /* The daemon port */
-    int port = CryptoNote::RPC_DEFAULT_PORT;
+    uint16_t port = CryptoNote::RPC_DEFAULT_PORT;
 
     /* The wallet file path */
     std::string walletFile;
